+++ src/pages/About/index.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import HolidazeCrew from "../../assets/images/holidazeCrew.jpg";
-import { PrimaryButton } from "../../components/Buttons/PrimaryButton";
+import { PrimaryButton } from "../../components/ui_elements/Buttons/PrimaryButton";
 
 export const About = () => {
   return (
